Added instrumentation for the beego HTTP client module instrumentation for different modules are separated to different packages
